fix(be): revert

+++ backend/apps/admin/src/contest/contest.service.ts
@@ -82,12 +82,15 @@ export class ContestService {
     groupId: number,
     contest: UpdateContestInput
   ): Promise<Contest> {
-    await this.prisma.contest.findFirstOrThrow({
+    const contestFound = await this.prisma.contest.findFirst({
       where: {
         id: contest.id,
         groupId
       }
     })
+    if (!contestFound) {
+      throw new EntityNotExistException('contest')
+    }
 
     if (contest.startTime >= contest.endTime) {
       throw new UnprocessableDataException(
@@ -113,18 +116,23 @@ export class ContestService {
   }
 
   async deleteContest(groupId: number, contestId: number) {
-    await this.prisma.contest.findFirstOrThrow({
+    const contest = await this.prisma.contest.findFirst({
       where: {
         id: contestId,
         groupId
       }
     })
+    if (!contest) {
+      throw new EntityNotExistException('contest')
+    }
 
-    return await this.prisma.contest.delete({
+    await this.prisma.contest.delete({
       where: {
         id: contestId
       }
     })
+
+    return contest
   }
 
   async getPublicizingRequests() {
@@ -197,12 +205,15 @@ export class ContestService {
       )
     }
 
-    const contest = await this.prisma.contest.findFirstOrThrow({
+    const contest = await this.prisma.contest.findFirst({
       where: {
         id: contestId,
         groupId
       }
     })
+    if (!contest) {
+      throw new EntityNotExistException('contest')
+    }
 
     let requests = (await this.cacheManager.get(
       PUBLICIZING_REQUEST_KEY
@@ -237,12 +248,15 @@ export class ContestService {
     contestId: number,
     problemIds: number[]
   ) {
-    const contest = await this.prisma.contest.findFirstOrThrow({
+    const contest = await this.prisma.contest.findUnique({
       where: {
         id: contestId,
         groupId
       }
     })
+    if (!contest) {
+      throw new EntityNotExistException('contest')
+    }
 
     const contestProblems: ContestProblem[] = []
